refactor(engine): moved governmentSuccession function from character.ts to politics.ts
