+++ src/engine/character.ts
@@ -5,6 +5,7 @@ import { CHARACTER_EVENTS } from '../data/events';
 
 //constants
 import { CYCLE_CONFIG } from '../constants/cycle_config';
+import { governmentSuccession } from './politics/politics';
 
 
 export function engineApplyCharacterProcess(currentState: GameState, charId: number, process: CharProcess): GameState {
@@ -219,72 +220,6 @@ export function engineUnassignCharacter(currentState: GameState, charId: number)
 			entities: newBuildings,
 		},
 	};
-}
-
-export function governmentSuccession(currentState: GameState, orgId: number): GameState {
-	let nextState = { ...currentState };
-	let functionOrg = { ...currentState.orgs.entities[orgId]};
-
-	if(functionOrg.government.succession === 'hereditary' && functionOrg.characters.leaderId){
-		const leader = nextState.characters.entities[functionOrg.characters.leaderId];
-
-		let nextLeader: Character | null = null;
-
-		for(const characterId of leader.history.childrenIds){
-			if(nextState.characters.entities[characterId].status === 'alive'){
-				nextLeader = { ...nextState.characters.entities[characterId]};
-				break;
-			}
-		}
-
-		if(!nextLeader){
-			nextLeader = generateCharacter(nextState.meta.lastCharacterId + 1, functionOrg.flavor.nameList);
-			nextLeader = {
-				...nextLeader,
-				citizenOrg: orgId,
-			}
-			nextState = { ...nextState, meta: { ...nextState.meta, lastCharacterId: nextState.meta.lastCharacterId + 1}, characters: { ids: [...nextState.characters.ids, nextLeader.id], entities: { ...nextState.characters.entities, [nextLeader.id]: nextLeader} }};
-		}
-
-		nextState = engineUnassignCharacter(nextState, functionOrg.characters.leaderId);
-
-		nextLeader = {
-			...nextLeader,
-			assignment: {
-				type: 'leader',
-				id: orgId,
-			}
-		};
-		functionOrg = {
-			...functionOrg,
-			characters: {
-				...functionOrg.characters,
-				leaderId: nextLeader.id
-			}
-		};
-
-		nextState = {
-			...nextState,
-			characters: {
-				...nextState.characters,
-				entities: {
-					...nextState.characters.entities,
-					[nextLeader.id]: nextLeader
-				}
-			},
-			orgs: {
-				...nextState.orgs,
-				entities: {
-					...nextState.orgs.entities,
-					[orgId]: functionOrg
-				}
-			}
-		}
-	}
-
-
-
-	return nextState;
 }
