+++ src/engine/politics/politics.ts
@@ -3,6 +3,79 @@ import { averagePopHappiness, findMigrationTarget, migratePop } from '../populat
 import { determinePotentialIdeology, spawnMovement } from './movements';
 import { spawnCellRandomLeader, selectNewCellAssignment } from './cells';
 import { CASSIGNMENT_CATALOG } from '../../data/cellActivities';
+import type { Character } from '../../types/charState';
+import { generateCharacter, engineUnassignCharacter } from '../character';
+
+
+
+export function governmentSuccession(currentState: GameState, orgId: number): GameState {
+    let nextState = { ...currentState };
+    let functionOrg = { ...currentState.orgs.entities[orgId]};
+
+    if(functionOrg.government.succession === 'hereditary' && functionOrg.characters.leaderId){
+        const leader = nextState.characters.entities[functionOrg.characters.leaderId];
+
+        let nextLeader: Character | null = null;
+
+        for(const characterId of leader.history.childrenIds){
+            if(nextState.characters.entities[characterId].status === 'alive'){
+                nextLeader = { ...nextState.characters.entities[characterId]};
+                break;
+            }
+        }
+
+        if(!nextLeader){
+            nextLeader = generateCharacter(nextState.meta.lastCharacterId + 1, functionOrg.flavor.nameList);
+            nextLeader = {
+                ...nextLeader,
+                citizenOrg: orgId,
+            }
+            nextState = { ...nextState, meta: { ...nextState.meta, lastCharacterId: nextState.meta.lastCharacterId + 1}, characters: { ids: [...nextState.characters.ids, nextLeader.id], entities: { ...nextState.characters.entities, [nextLeader.id]: nextLeader} }};
+        }
+
+        nextState = engineUnassignCharacter(nextState, functionOrg.characters.leaderId);
+
+        nextLeader = {
+            ...nextLeader,
+            assignment: {
+                type: 'leader',
+                id: orgId,
+            }
+        };
+        functionOrg = {
+            ...functionOrg,
+            characters: {
+                ...functionOrg.characters,
+                leaderId: nextLeader.id
+            }
+        };
+
+        nextState = {
+            ...nextState,
+            characters: {
+                ...nextState.characters,
+                entities: {
+                    ...nextState.characters.entities,
+                    [nextLeader.id]: nextLeader
+                }
+            },
+            orgs: {
+                ...nextState.orgs,
+                entities: {
+                    ...nextState.orgs.entities,
+                    [orgId]: functionOrg
+                }
+            }
+        }
+    }
+
+
+
+    return nextState;
+}
+
+
+
 
 export function processPolitics(currentState: GameState ): EngineResult {
     const planetoids = Object.values(currentState.planetoids.entities);
